Split DistrhoUIInternal base struct into separate file; cleanup
