+++ distrho/src/DistrhoUIInternal.hpp
@@ -17,7 +17,7 @@
 #ifndef DISTRHO_UI_INTERNAL_HPP_INCLUDED
 #define DISTRHO_UI_INTERNAL_HPP_INCLUDED
 
-#include "../DistrhoUI.hpp"
+#include "DistrhoUIPrivateData.hpp"
 
 #if DISTRHO_PLUGIN_HAS_EXTERNAL_UI
 # include "../extra/Sleep.hpp"
@@ -46,149 +46,17 @@ extern Window*     d_lastUiWindow;
 #endif
 
 // -----------------------------------------------------------------------
-// UI callbacks
 
-typedef void (*editParamFunc)   (void* ptr, uint32_t rindex, bool started);
-typedef void (*setParamFunc)    (void* ptr, uint32_t rindex, float value);
-typedef void (*setStateFunc)    (void* ptr, const char* key, const char* value);
-typedef void (*sendNoteFunc)    (void* ptr, uint8_t channel, uint8_t note, uint8_t velo);
-typedef void (*setSizeFunc)     (void* ptr, uint width, uint height);
-typedef bool (*fileRequestFunc) (void* ptr, const char* key);
-
-// -----------------------------------------------------------------------
-// UI private data
-
-struct UI::PrivateData {
-    // DSP
-    double   sampleRate;
-    uint32_t parameterOffset;
-#if DISTRHO_PLUGIN_WANT_DIRECT_ACCESS
-    void*    dspPtr;
+#if DISTRHO_PLUGIN_HAS_EXTERNAL_UI
+UI* createUiWrapper(void* dspPtr, uintptr_t winId, double scaleFactor, const char* bundlePath);
+#else // DISTRHO_PLUGIN_HAS_EXTERNAL_UI
+UI* createUiWrapper(void* dspPtr, Window* window);
 #endif
-
-    // UI
-    bool automaticallyScale;
-    bool resizeInProgress;
-    uint minWidth;
-    uint minHeight;
-
-    // Callbacks
-    void*           callbacksPtr;
-    editParamFunc   editParamCallbackFunc;
-    setParamFunc    setParamCallbackFunc;
-    setStateFunc    setStateCallbackFunc;
-    sendNoteFunc    sendNoteCallbackFunc;
-    setSizeFunc     setSizeCallbackFunc;
-    fileRequestFunc fileRequestCallbackFunc;
-
-    PrivateData() noexcept
-        : sampleRate(d_lastUiSampleRate),
-          parameterOffset(0),
-#if DISTRHO_PLUGIN_WANT_DIRECT_ACCESS
-          dspPtr(d_lastUiDspPtr),
-#endif
-          automaticallyScale(false),
-          resizeInProgress(false),
-          minWidth(0),
-          minHeight(0),
-          callbacksPtr(nullptr),
-          editParamCallbackFunc(nullptr),
-          setParamCallbackFunc(nullptr),
-          setStateCallbackFunc(nullptr),
-          sendNoteCallbackFunc(nullptr),
-          setSizeCallbackFunc(nullptr),
-          fileRequestCallbackFunc(nullptr)
-    {
-        DISTRHO_SAFE_ASSERT(d_isNotZero(sampleRate));
-
-#if defined(DISTRHO_PLUGIN_TARGET_DSSI) || defined(DISTRHO_PLUGIN_TARGET_LV2)
-        parameterOffset += DISTRHO_PLUGIN_NUM_INPUTS + DISTRHO_PLUGIN_NUM_OUTPUTS;
-# if DISTRHO_PLUGIN_WANT_LATENCY
-        parameterOffset += 1;
-# endif
-#endif
-
-#ifdef DISTRHO_PLUGIN_TARGET_LV2
-# if (DISTRHO_PLUGIN_IS_SYNTH || DISTRHO_PLUGIN_WANT_TIMEPOS || DISTRHO_PLUGIN_WANT_STATE)
-        parameterOffset += 1;
-#  if DISTRHO_PLUGIN_WANT_STATE
-        parameterOffset += 1;
-#  endif
-# endif
-#endif
-    }
-
-    void editParamCallback(const uint32_t rindex, const bool started)
-    {
-        if (editParamCallbackFunc != nullptr)
-            editParamCallbackFunc(callbacksPtr, rindex, started);
-    }
-
-    void setParamCallback(const uint32_t rindex, const float value)
-    {
-        if (setParamCallbackFunc != nullptr)
-            setParamCallbackFunc(callbacksPtr, rindex, value);
-    }
-
-    void setStateCallback(const char* const key, const char* const value)
-    {
-        if (setStateCallbackFunc != nullptr)
-            setStateCallbackFunc(callbacksPtr, key, value);
-    }
-
-    void sendNoteCallback(const uint8_t channel, const uint8_t note, const uint8_t velocity)
-    {
-        if (sendNoteCallbackFunc != nullptr)
-            sendNoteCallbackFunc(callbacksPtr, channel, note, velocity);
-    }
-
-    void setSizeCallback(const uint width, const uint height)
-    {
-        if (setSizeCallbackFunc != nullptr)
-            setSizeCallbackFunc(callbacksPtr, width, height);
-    }
-
-    bool fileRequestCallback(const char* key)
-    {
-        if (fileRequestCallbackFunc != nullptr)
-            return fileRequestCallbackFunc(callbacksPtr, key);
-
-        // TODO use old style DPF dialog here
-
-        return false;
-    }
-};
 
 // -----------------------------------------------------------------------
 // Plugin Window, needed to take care of resize properly
 
-#if DISTRHO_PLUGIN_HAS_EXTERNAL_UI
-static inline
-UI* createUiWrapper(void* const dspPtr, const uintptr_t winId, const double scaleFactor, const char* const bundlePath)
-{
-    d_lastUiDspPtr    = dspPtr;
-    g_nextWindowId    = winId;
-    g_nextScaleFactor = scaleFactor;
-    g_nextBundlePath  = bundlePath;
-    UI* const ret     = createUI();
-    d_lastUiDspPtr    = nullptr;
-    g_nextWindowId    = 0;
-    g_nextScaleFactor = 1.0;
-    g_nextBundlePath  = nullptr;
-    return ret;
-}
-#else // DISTRHO_PLUGIN_HAS_EXTERNAL_UI
-static inline
-UI* createUiWrapper(void* const dspPtr, Window* const window)
-{
-    d_lastUiDspPtr = dspPtr;
-    d_lastUiWindow = window;
-    UI* const ret  = createUI();
-    d_lastUiDspPtr = nullptr;
-    d_lastUiWindow = nullptr;
-    return ret;
-}
-
+#if ! DISTRHO_PLUGIN_HAS_EXTERNAL_UI
 class UIExporterWindow : public Window
 {
 public:
@@ -278,7 +146,7 @@ public:
 #if DISTRHO_PLUGIN_HAS_EXTERNAL_UI
         : fUI(createUiWrapper(dspPtr, winId, scaleFactor, bundlePath)),
 #else
-        : glApp(),
+        : glApp(DISTRHO_UI_IS_STANDALONE),
           glWindow(glApp, winId, scaleFactor, dspPtr),
           fChangingSize(false),
           fUI(glWindow.getUI()),
